[ Widget Preview ] Move `preview_detector_test.dart` from `general.shard` to `commands.shard` (#163619)
Tests under the `general.shard` have a 2000ms timeout and these tests
write to the real file system and watch directories for changes. This
can be slow on heavily loaded machines and cause flaky failures.
